requirements.txt
# Core AI & API
openai
python-dotenv

# Vector Search & Math
numpy
faiss-cpu  # Use faiss-gpu if you have a supported NVIDIA GPU
scipy

# Document & Image Processing
PyPDF2
# Note: Ensure you have any specific dependencies for your custom image_loader (like Pillow or pytesseract)

# Audio & Voice
sounddevice
pygame

# Utilities (often needed for RAG pipelines)
tiktoken